Cleanup Gradle Code
