--- cdtool/cdgradle/src/main/java/de/monticore/cdgen/gradleplugin/CDGenAction.java
@@ -1,19 +1,0 @@
-/* (c) https://github.com/MontiCore/monticore */
-package de.monticore.cdgen.gradleplugin;
-
-import de.monticore.gradle.common.AToolAction;
-import de.monticore.gradle.internal.isolation.CachedIsolation;
-
-public abstract class CDGenAction extends AToolAction {
-  
-  protected static final CachedIsolation.WithClassPath isolator =
-      new CachedIsolation.WithClassPath();
-  
-  @Override
-  protected void doRun(String[] args) {
-    final String prefix = "[" + getParameters().getProgressName().get() + "] ";
-    isolator.executeInClassloader(CDGenToolInvoker.class.getName(), "run", args, prefix,
-        getParameters().getExtraClasspathElements());
-  }
-  
-}
