+++ cdtool/cdgradle/src/main/java/de/monticore/cdgen/gradleplugin/CDGenGradlePlugin.java
@@ -1,18 +1,16 @@
 /* (c) https://github.com/MontiCore/monticore */
 package de.monticore.cdgen.gradleplugin;
 
-import java.io.IOException;
-import java.util.Properties;
 import java.util.stream.Collectors;
+
+import de.monticore.gradle.queue.CachedQueueServicePlugin;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
-import org.gradle.api.artifacts.Configuration;
 import org.gradle.api.file.Directory;
 import org.gradle.api.file.FileCollection;
 import org.gradle.api.file.SourceDirectorySet;
 import org.gradle.api.internal.lambdas.SerializableLambdas;
 import org.gradle.api.model.ObjectFactory;
-import org.gradle.api.plugins.JavaLibraryPlugin;
 import org.gradle.api.plugins.JavaPluginExtension;
 import org.gradle.api.provider.Provider;
 import org.gradle.api.tasks.SourceSet;
@@ -22,11 +20,6 @@ import javax.inject.Inject;
 @SuppressWarnings("unused")
 public class CDGenGradlePlugin implements Plugin<Project> {
   
-  /**
-   * Configuration containing the classpath of the generator tool.
-   */
-  public static final String CONFIG_TOOL = "cdTool";
-  
   private final ObjectFactory objectFactory;
   
   @Inject
@@ -34,43 +27,14 @@ public class CDGenGradlePlugin implements Plugin<Project> {
     this.objectFactory = objectFactory;
   }
   
-  /**
-   * Configuration containing the classpath of the target.
-   * it will be added to the api configuration & passed as a symbol path for class2mc
-   */
-  public static final String CONFIG_TARGET_RUNTIME = "cdToolTargetRuntime";
-  
   @Override
   public void apply(Project project) {
-    project.getPluginManager().apply(JavaLibraryPlugin.class);
+    // Set up the improved work-queue
+    project.getPluginManager().apply(CachedQueueServicePlugin.class);
     
-    // Setup cdTool dependency
-    var properties = loadProperties();
-    String version = properties.getProperty("version");
-    
-    Configuration toolConfig = project.getConfigurations().maybeCreate(CONFIG_TOOL);
-    toolConfig.setCanBeResolved(true);
-    Configuration toolRuntimeConfig = project.getConfigurations().maybeCreate(
-        CONFIG_TARGET_RUNTIME);
-    toolRuntimeConfig.setCanBeResolved(true);
-    
-    toolConfig.defaultDependencies(dependencies -> {
-      dependencies.add(project.getDependencies().create("de.monticore.lang:cd4analysis:"
-          + version));
-    });
-    
-    toolRuntimeConfig.defaultDependencies(dependencies -> {
-      dependencies.add(project.getDependencies().create("de.monticore.lang:cd-runtime:" + version
-          + ":cd-runtime"));
-    });
-    
-    project.getTasks().withType(CDGenTask.class).configureEach(t -> t.getExtraClasspathElements()
-        .from(toolConfig));
-    
-    project.getTasks().withType(CDGenTask.class).configureEach(t -> t.getTargetSymbolPath().from(
-        toolRuntimeConfig));
-    
-    project.getConfigurations().named("api").configure(api -> api.extendsFrom(toolRuntimeConfig));
+    // Populate the "cdTool" configuration with the generator itself and
+    // Populate the "cdToolTargetRuntime" configuration with the runtime
+    project.getPluginManager().apply(CDGenGradleDependenciesPlugin.class);
     
     // Set up source-Sets
     project.getExtensions().getByType(JavaPluginExtension.class).getSourceSets().all(sourceSet -> {
@@ -98,17 +62,6 @@ public class CDGenGradlePlugin implements Plugin<Project> {
       project.getTasks().named(sourceSet.getCompileJavaTaskName()).configure(t -> t.dependsOn(
           task));
     });
-  }
-  
-  public Properties loadProperties() {
-    Properties properties = new Properties();
-    try {
-      properties.load(this.getClass().getClassLoader().getResourceAsStream("buildInfo.properties"));
-    }
-    catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-    return properties;
   }
   
   /** Adds the "cd" extension to every source set */
